Refactor code: fix styles
