+++ src/pages/SignUp.jsx
@@ -176,7 +176,7 @@ export default function SignUp() {
             }}
           >
             <Avatar sx={{ m: 1, backgroundColor: 'primary.main' }}>
-              <LockOutlinedIcon sx={{color: customColors.white}}/>
+              <LockOutlinedIcon sx={{ color: customColors.white }} />
             </Avatar>
             <Typography component="h1" variant="h5">
               Sign up
